requirements.txt
# Librerías de trading y datos
yfinance==0.2.31
pandas==2.1.4
numpy==1.26.2
oandapyV20==0.7.1
apscheduler==3.10.4

# Librerías de análisis
ta-lib==0.4.20
scikit-learn==1.3.2

# Logging y configuración
python-dotenv==1.0.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0

# Documentación
sphinx==7.2.6

# Deployment
gunicorn==21.2.0
